+++ write.go
@@ -45,11 +45,18 @@ func (s SrtSocket) Write(b []byte) (n int, err error) {
 	}
 	n, err = srtSendMsg2Impl(s.socket, b, nil)
 
+	// Issue 3
+	// Previously, this for loop would never break if the deadline has been hit (as per return value of s.pd.wait(ModeWrite))
 	for {
 		if !errors.Is(err, error(EAsyncSND)) || s.blocking {
 			return
 		}
-		s.pd.wait(ModeWrite)
+
+		// E.G., Error because we reached timed out. Like we do for connect.
+		if err = s.pd.wait(ModeWrite); err != nil {
+			return
+		}
+
 		n, err = srtSendMsg2Impl(s.socket, b, nil)
 	}
 }