+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - [SA-1463] Return error when trying to create a new socket
 - [SA-2074] Add a ReadPacket function
+- Add various fixes for SetDeadline functionality